Move socks config into a sparate folder
